requirements.txt
streamlit
python-pptx
langchain
langchain-groq
langchain-community
langchain-text-splitters
pydantic
python-dotenv
langchain-core